Refactor element validation functions into separate utility module
Move validation functions (isReadableTag, hasMinimumTextContent,
isElementVisible, shouldShowPlayButton, isReadableElement,
filterReadableElements) from readableElements.js config to new
elementValidation.js utility module. This improves separation of
concerns by keeping configuration separate from validation logic.

- Extract validation functions to plugin/src/shared/utils/elementValidation.js
- Update imports in content/index.js and content/ui/PlayButton.js
- Simplify readableElements.js to focus on configuration only
- Update test imports and add comprehensive tests for new module
- Add global console mocks to test setup for cleaner output
