+++ plugin/src/content/index.js
@@ -11,7 +11,8 @@ import { PLAYER_STATES } from '../shared/utils/constants.js';
 import { ParagraphQueue } from './queue/ParagraphQueue.js';
 import { PrefetchManager } from './prefetch/PrefetchManager.js';
 import { ContinuousPlaybackController } from './controllers/ContinuousPlaybackController.js';
-import { getReadableElementsSelector, filterReadableElements } from '../shared/config/readableElements.js';
+import { getReadableElementsSelector } from '../shared/config/readableElements.js';
+import { filterReadableElements } from '../shared/utils/elementValidation.js';
 
 class TTSContentScript {
   constructor() {
