+++ plugin/tests/setup.js
@@ -1,4 +1,18 @@
-import { vi } from 'vitest';
+import { vi, beforeEach, afterEach } from 'vitest';
+
+// Mock console methods globally to keep test output clean
+beforeEach(() => {
+  global.consoleMocks = {
+    error: vi.spyOn(console, 'error').mockImplementation(() => {}),
+    warn: vi.spyOn(console, 'warn').mockImplementation(() => {}),
+    log: vi.spyOn(console, 'log').mockImplementation(() => {}),
+    info: vi.spyOn(console, 'info').mockImplementation(() => {}),
+  };
+});
+
+afterEach(() => {
+  vi.restoreAllMocks();
+});
 
 // Setup global mocks
 global.chrome = {
